hw_conf: Added possibility to check HW limits
The Hw limit shall be defined in the target specific hwconf file.
Each hw limit shall consists of a min and a max value.
The hw check will run on system start and each time new hw parameters
are written. Values outside of the specified range will be saturated to
min or max value.

Change-Id: I461282377585553b715a1f0f2c364dcc5665eb15