datatypes: Removed struct packing.
There is a high risk of unaligned access hard faults if the backup data
struct and main config structs are packed.
Operating on floats through pointer in the structs will cause
unaligned access fault if the float is unaligned which currently a lot
of them are.
By removing the packed directive we increase the struct size from
374 bytes to 400 bytes this is however still well below the backup flash
area of 8kb.
Also, with this change no padding should be needed in the backup data
structure.

Change-Id: Ibcbb8526b45c7df9e3eaf567d7c9ce623498d5f1