hw_rbat: Change scaling for HW_GET_V_MOIST_SENSE.
The measured voltage at the water detection pin is added to the
send buffer as a float16 with a scaling factor of 1e4.
When adding a float16 to the buffer, it will be converted to a int16.
So if the measured voltage is greater than 3.3V this will result in
an overflow.
By decreasing the scaling to 1e3 we will prevent the overflow.